学霸学渣SFT+DPO/requirements_training.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.8.0+cu128
transformers==4.57.6
trl==0.24.0
peft==0.17.1
datasets==4.5.0
accelerate==1.10.1
bitsandbytes==0.48.2
numpy==1.26.4
PyYAML==6.0.2
